termionix_ansicodec::ansi

Enum AnsiControlSequenceIntroducer

pub enum AnsiControlSequenceIntroducer {

    CursorUp(u8),
    CursorDown(u8),
    CursorForward(u8),
    CursorBack(u8),
    CursorNextLine(u8),
    CursorPreviousLine(u8),
    CursorHorizontalAbsolute(u8),
    CursorPosition {
        row: u8,
        col: u8,
    },
    DeviceStatusReport,
    SaveCursorPosition,
    RestoreCursorPosition,
    EraseInDisplay(EraseInDisplayMode),
    EraseInLine(EraseInLineMode),
    SetMode,
    ResetMode,
    DECPrivateModeSet,
    DECPrivateModeReset,
    ScrollUp,
    ScrollDown,
    InsertCharacter,
    DeleteCharacter,
    InsertLine,
    DeleteLine,
    EraseCharacter,
    TextCursorEnableMode,
    AlternativeScreenBuffer,
    SetKeyboardStrings,
    Unknown,

}

Control Sequence Introducer (CSI) Command

Variants

CursorUp(u8)

CUU - Cursor Up ESC[#A - moves cursor up # lines

CursorDown(u8)

CUD - Cursor Down ESC[#B - moves cursor down # lines

CursorForward(u8)

CUF - Cursor Forward ESC[#C - moves cursor right # columns

CursorBack(u8)

CUB - Cursor Back ESC[#D - moves cursor left # columns

CursorNextLine(u8)

CNL - Cursor Next Line ESC[#E - moves the cursor to the beginning of the next line, # lines down

CursorPreviousLine(u8)

CPL - Cursor Previous Line ESC[#F - moves the cursor to the beginning of the previous line, # lines up

CursorHorizontalAbsolute(u8)

CHA - Cursor Horizontal Absolute ESC[#G - moves cursor to column #

CursorPosition

CUP - Cursor Position / HVP - Horizontal Vertical Position ESC[{line};{column}H or ESC[{line};{column}f

Fields

row: u8

Cursor Row

col: u8

Cursor Column

DeviceStatusReport

DSR - Device Status Report ESC[6n - request cursor position (reports as ESC[#;#R)

SaveCursorPosition

SCP - Save Cursor Position (SCO) ESC[s

RestoreCursorPosition

RCP - Restore Cursor Position (SCO) ESC[u

EraseInDisplay(EraseInDisplayMode)

ED - Erase in Display ESC[J or ESC[0J - erase from cursor until end of screen ESC[1J - erase from cursor to beginning of screen ESC[2J - erase entire screen ESC[3J - erase saved lines

EraseInLine(EraseInLineMode)

EL - Erase in Line ESC[K or ESC[0K - erase from cursor to end of line ESC[1K - erase start of line to the cursor ESC[2K - erase the entire line

SetMode

SM - Set Mode ESC[={value}h - Changes screen width or type

ResetMode

RM - Reset Mode ESC[={value}l - Resets the mode

DECPrivateModeSet

DECSET - DEC Private Mode Set ESC[?{value}h - DEC private mode set

DECPrivateModeReset

DECRST - DEC Private Mode Reset ESC[?{value}l - DEC private mode reset

ScrollUp

SU - Scroll Up ESC[#S - Scroll up # lines

ScrollDown

SD - Scroll Down ESC[#T - Scroll down # lines

InsertCharacter

ICH - Insert Character ESC[#@ - Insert # blank characters

DeleteCharacter

DCH - Delete Character ESC[#P - Delete # characters

InsertLine

IL - Insert Line ESC[#L - Insert # blank lines

DeleteLine

DL - Delete Line ESC[#M - Delete # lines

EraseCharacter

ECH - Erase Character ESC[#X - Erase # characters from the cursor position

TextCursorEnableMode

DECTCEM - Text Cursor Enable Mode ESC[?25h - Show cursor ESC[?25l - Hide cursor

AlternativeScreenBuffer

Alt Screen - Alternative Screen Buffer ESC[?1049h - Enable alternative buffer ESC[?1049l - Disable alternative buffer

SetKeyboardStrings

Set Keyboard Strings ESC[{code};{string};{…}p

Unknown

Unknown or unsupported CSI command

Implementations

impl AnsiControlSequenceIntroducer

pub fn len(&self) -> usize

Returns the encoded byte length of this CSI command.

This method calculates the total number of bytes produced when the CSI command is encoded to its escape sequence representation. The length includes the ESC [ introducer (2 bytes), all parameter bytes, and the final command byte.

Returns

The number of bytes in the encoded form. Minimum is 3 bytes for simple commands (e.g., ESC [ A), up to 10+ bytes for commands with multiple parameters.

Performance

This is O(1) for most commands. Some parameterized commands require digit counting which is O(log n) where n is the parameter value.

Examples

use termionix_ansicodec::ansi::AnsiControlSequenceIntroducer;

assert_eq!(AnsiControlSequenceIntroducer::CursorUp(1).len(), 4); // ESC[1A
assert_eq!(AnsiControlSequenceIntroducer::CursorUp(255).len(), 6); // ESC[255A
assert_eq!(AnsiControlSequenceIntroducer::SaveCursorPosition.len(), 3); // ESC[s

Use Cases

Buffer allocation: Pre-allocate space before encoding

Terminal control: Calculate message sizes for terminal commands

pub fn encode<T: BufMut>(&self, dst: &mut T) -> AnsiCodecResult<usize>

Encode this CSI command to a BufMut buffer.

This method encodes the CSI command into its escape sequence representation and writes it to the provided mutable buffer. The buffer’s write position is automatically advanced.

Arguments

dst - A mutable buffer implementing BufMut

Returns

Returns Ok(bytes_written) on success, containing the number of bytes written. Returns an AnsiResult error if encoding fails.

CSI Format

All CSI commands are encoded as ESC [ <parameters> <final_byte>:

ESC: 0x1B

[: 0x5B

Parameters: numeric values and semicolons (variable length)

Final byte: command identifier (0x40-0x7E)

Examples

use bytes::BytesMut;
use termionix_ansicodec::ansi::AnsiControlSequenceIntroducer;

let cmd = AnsiControlSequenceIntroducer::CursorUp(5);
let mut buffer = BytesMut::new();
let written = cmd.encode(&mut buffer).unwrap();
assert_eq!(written, 4); // ESC[5A

See Also

write() - Write to a std::io::Write trait object

pub fn write<W: Write>(&self, writer: &mut W) -> Result<usize>

Write this CSI command to a std::io::Write writer.

This method performs the actual encoding of the CSI command into its escape sequence representation and writes it to the provided writer.

Arguments

writer - A mutable writer implementing std::io::Write

Returns

Returns Ok(bytes_written) on success, containing the number of bytes written. Returns std::io::Error if writing fails.

Encoding Behavior

Each command produces its specific escape sequence:

CursorUp(n): ESC [ n A

CursorDown(n): ESC [ n B

CursorForward(n): ESC [ n C

CursorBack(n): ESC [ n D

CursorPosition { row, col }: ESC [ row ; col H

EraseInDisplay(mode): ESC [ mode J

EraseInLine(mode): ESC [ mode K

SaveCursorPosition: ESC [ s

RestoreCursorPosition: ESC [ u

And others as documented in AnsiControlSequenceIntroducer variants

Examples

use termionix_ansicodec::ansi::AnsiControlSequenceIntroducer;

let cmd = AnsiControlSequenceIntroducer::SaveCursorPosition;
let mut output = Vec::new();
let written = cmd.write(&mut output).unwrap();
assert_eq!(written, 3); // ESC[s
assert_eq!(output, b"\x1b[s");

See Also

encode() - Encode to a BufMut buffer

len() - Get the encoded byte length